+++ pubspec.yaml
@@ -16,6 +16,7 @@ dependencies:
   path_provider: ^0.4.0
   shared_preferences: ^0.4.1
   http: ^0.11.3
+  fluro: ^1.3.1
 
 
 dev_dependencies:
